+++ CHANGES.rst
@@ -46,6 +46,7 @@ Development
 - [doc] Fix building the documentation
 - [mw] Fix logging. Thanks, @think-nice-things!
 - [cli] Start using the ``Click`` command line parser
+- [mw] Improve "minimal" application framework bootstrapping
 
 
 2019-11-01 0.169.3